General delete confirmation added, improved user and role CRUD
Also included:
- Breadcrumbs rendering improved: no extra spaces, using default icons
- Currently logged in user can not be deleted
- UI standardization: removed delete buttons from edit pages
- UI standardization: added delete buttons to show pages
- UI standardization: create/edit from actions are shown in card-footer
- Added further tricks to prevent aggressive browser auto-completion

+++ src/Http/Controllers/UserController.php
@@ -12,6 +12,7 @@
 
 namespace Konekt\AppShell\Http\Controllers;
 
+use Illuminate\Support\Facades\Auth;
 use Konekt\Acl\Models\RoleProxy;
 use Konekt\AppShell\Contracts\Requests\CreateUser;
 use Konekt\AppShell\Contracts\Requests\UpdateUser;
@@ -116,8 +117,7 @@ class UserController extends BaseController
             return redirect()->back();
         }
 
-        //@todo process route prefixes based on box config
-        return redirect(route('appshell.user.index'));
+        return redirect(route('appshell.user.show', $user));
     }
 
     /**
@@ -129,6 +129,12 @@ class UserController extends BaseController
      */
     public function destroy(User $user)
     {
+        if ($user->id == Auth::user()->id) {
+            flash()->error(__("You can't delete your self. What's wrong dude? Do you want to talk about it?"));
+
+            return redirect()->back();
+        }
+
         try {
             $user->delete();
 
@@ -137,7 +143,6 @@ class UserController extends BaseController
             flash()->error(__('Error: :msg', ['msg' => $e->getMessage()]));
         }
 
-        //@todo process route prefixes based on box config
         return redirect(route('appshell.user.index'));
     }
